requirements.txt
numpy
pandas
scikit-learn
mlflow==3.8.1
fastapi
uvicorn
evidently
prometheus-client
prometheus-fastapi-instrumentator
pytest
python-dotenv
